+++ CHANGELOG.md
@@ -11,6 +11,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## Changed
 - Moved system command modules into their own area
+- Rename ComponentNameUtil to ComponentNameExtractor and moved to common
 
 ### [6.0.0-M28] - 2019-06-05
 ## Added